Move nested tests to their own separate test files
